Finish date formatting.
